+++ src/lib-components/Flexible/Info.vue
