+++ src/lib-components/index.js
@@ -50,7 +50,6 @@ export { default as DividerWayFinder } from './DividerWayFinder.vue'
 export { default as FiltersDropdown } from './FiltersDropdown.vue'
 export { default as FlexibleAssociatedTopicCards } from './Flexible/AssociatedTopicCards.vue'
 export { default as FlexibleBannerFeatured } from './Flexible/BannerFeatured.vue'
-export { default as FlexibleBlockInfo } from './Flexible/BlockInfo.vue'
 export { default as FlexibleBlocks } from './FlexibleBlocks.vue'
 export { default as FlexibleCallToAction } from './Flexible/CallToAction.vue'
 export { default as FlexibleCardWithImage } from './Flexible/CardWithImage.vue'
@@ -60,6 +59,7 @@ export { default as FlexibleHighlight } from './Flexible/Highlight.vue'
 export { default as FlexibleHorizontalDivider } from './Flexible/HorizontalDivider.vue'
 export { default as FlexibleImpactNumberCards } from './Flexible/ImpactNumberCards.vue'
 export { default as FlexibleImpactNumbersCarousel } from './Flexible/ImpactNumbersCarousel.vue'
+export { default as FlexibleInfo } from './Flexible/Info.vue'
 export { default as FlexibleMediaGallery } from './Flexible/MediaGallery.vue'
 export { default as FlexibleMediaGalleryBannerImage } from './Flexible/MediaGallery/BannerImage.vue'
 export { default as FlexibleMediaGalleryNewLightbox } from './Flexible/MediaGallery/NewLightbox.vue'
